slight refactor code still works
